requirements.txt
fastapi
uvicorn[standard]

sqlalchemy
psycopg[binary]

python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

pydantic
pydantic-settings
email-validator

python-multipart
python-dotenv

redis
websockets

alembic

psycopg2-binary==2.9.9
gunicorn==21.2.0